feat: Add ApiError & Location

+++ ewm-server/src/main/java/ru/practicum/compilation/CompilationMainService.java
@@ -6,11 +6,13 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
-import org.springframework.web.server.ResponseStatusException;
+import org.springframework.transaction.annotation.Transactional;
 import ru.practicum.compilation.dto.CompilationDto;
 import ru.practicum.compilation.dto.CompilationMapper;
 import ru.practicum.compilation.repository.CompilationRepository;
 import ru.practicum.error.ApiError;
+import ru.practicum.event.Event;
+import ru.practicum.event.repository.EventRepository;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -23,7 +25,10 @@ import static java.util.Optional.of;
 @Slf4j
 public class CompilationMainService implements CompilationService {
     private final CompilationRepository compilationRepository;
+    private final CompilationMapper compilationMapper;
+    private final EventRepository eventRepository;
 
+    @Transactional(readOnly = true)
     @Override
     public Collection<CompilationDto> getCompilations(Boolean pinned, Integer from, Integer size) {
         PageRequest pageRequest = PageRequest.of(this.getPageNumber(from, size), size,
@@ -31,43 +36,69 @@ public class CompilationMainService implements CompilationService {
         Iterable<Compilation> compilationsPage = compilationRepository.findCompilationsByPinned(pinned, pageRequest);
         Collection<Compilation> compilations = new ArrayList<>();
         compilationsPage.forEach(compilations::add);
-        return CompilationMapper.toCompilationDto(compilations);
+        return compilationMapper.toCompilationDto(compilations);
     }
 
+    @Transactional(readOnly = true)
     @Override
     public Optional<CompilationDto> getCompilationById(Long compilationId) {
-        return of(CompilationMapper.toCompilationDto(getById(compilationId)));
+        return of(compilationMapper.toCompilationDto(getById(compilationId)));
     }
 
+    @Transactional
     @Override
-    public Optional<CompilationDto> addCompilation(Compilation compilation) {
-        return of(CompilationMapper.toCompilationDto(compilationRepository.save(compilation)));
+    public Optional<CompilationDto> addCompilation(CompilationDto compilation) {
+        return of(compilationMapper.toCompilationDto(compilationRepository
+                .save(compilationMapper
+                        .fromCompilationDto(compilation))));
     }
 
+    @Transactional
     @Override
     public void deleteCompilation(Long compilationId) {
-        compilationRepository.delete(getById(compilationId));
+        if (getById(compilationId).getEvents().isEmpty()) {
+            compilationRepository.delete(getById(compilationId));
+        } else {
+            throw new ApiError(HttpStatus.BAD_REQUEST, "При удалении подборки",
+                    "Подборка содержит события");
+        }
     }
 
+    @Transactional
     @Override
     public boolean removeEventFromCompilation(Long compId, Long eventId) {
-        return getById(compId).getEvents_in().remove(eventId);
+        return getById(compId).getEvents().remove(eventRepository.findById(eventId).orElseThrow(
+                () -> new ApiError(HttpStatus.NOT_FOUND, "Не найдено событие",
+                        String.format("Не найдено событие с id %s при попытке добавления в подборку", eventId)
+                )));
     }
 
+    @Transactional
     @Override
     public boolean addEventToCompilation(Long compId, Long eventId) {
-        return getById(compId).getEvents_in().add(eventId);
+        return getById(compId).getEvents().add(eventRepository.findById(eventId).orElseThrow(
+                () -> new ApiError(HttpStatus.NOT_FOUND, "Не найдено событие",
+                        String.format("Не найдено событие с id %s при попытке добавления в подборку", eventId)
+        )));
     }
 
+    @Transactional
     @Override
     public boolean removeCompilationFromHomePage(Long compId) {
-        getById(compId).setPinned(Boolean.FALSE);
+        compilationRepository.findById(compId).orElseThrow(
+                () -> new ApiError(HttpStatus.NOT_FOUND, "Подборка не найдена",
+                        String.format("Подборка с id %s не найдена", compId))
+        ).setPinned(Boolean.FALSE);
         return true;
     }
 
+    @Transactional
     @Override
     public boolean addCompilationToHomePage(Long compId) {
-        getById(compId).setPinned(Boolean.TRUE);
+        compilationRepository.findById(compId).orElseThrow(
+                () -> new ApiError(HttpStatus.NOT_FOUND, "Подборка не найдена",
+                        String.format("Подборка с id %s не найдена", compId))
+        ).setPinned(Boolean.TRUE);
         return true;
     }
 
@@ -81,5 +112,4 @@ public class CompilationMainService implements CompilationService {
     private Integer getPageNumber(Integer from, Integer size) {
         return from % size;
     }
-
 }
